backend/src/main/java/com/github/chuettenrauch/mixifyapi/user/service/UserService.java
package com.github.chuettenrauch.mixifyapi.user.service;

import com.github.chuettenrauch.mixifyapi.user.model.User;
import com.github.chuettenrauch.mixifyapi.user.repository.UserRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
@RequiredArgsConstructor
public class UserService {

    private final UserRepository userRepository;

    public Optional<User> findByEmail(String email) {
        return this.userRepository.findByEmail(email);
    }

}
